+++ sdk/communication/communication-job-router/src/models/requests.ts
@@ -1,11 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-
-
 export {
   CancelJobRequest,
   CloseJobRequest,
-  CompleteJobRequest,
-  ReleaseAssignmentRequest,
+  CompleteJobRequest
 } from "../generated/src/models";