+++ sdk/communication/communication-job-router/CHANGELOG.md
@@ -7,6 +7,6 @@ The first preview of the Azure Communication Job Router Client has the following
 - create/get/update/delete/list (a) classification policy(s)
 - create/get/update/delete/list (a) distribution policy(s)
 - create/get/update/delete/list (a) exception policy(s)
-- create/get/update/delete/list/cancel/complete/close (a) job(s)
+- create/get/update/delete/list/cancel/complete/close/reclassify/updateClassification (a) job(s)
 - create/get/update/delete/list (a) queue(s)
-- register/get/deregister/delete/list/release assignment (a) worker(s)
+- create/register/get/deregister/delete/list (a) worker(s)